+++ src/main/java/com/luckytree/shop_service/common/config/FeignClientConfig.java
@@ -1,4 +1,4 @@
-package com.luckytree.shop_service.config;
+package com.luckytree.shop_service.common.config;
 
 import feign.Logger;
 import lombok.extern.java.Log;
